requirements.txt
discord.py[voice]
wavelink==1.3.3
asyncio